% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotGeneNames.R
\docType{data}
\name{annotGeneNames}
\alias{annotGeneNames}
\title{gene names of probes in 450K array dat}
\format{A vector with length 480457}
\description{
gene names of probes in 450K array dat
}
\keyword{datasets}
