% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPM-package.R
\docType{package}
\name{BPM}
\alias{BPM}
\alias{BPM-package}
\title{BPM software package}
\description{
Bayesian model for purity estimation using DNA
methylation data
}
\details{
The main function is \code{\link{BayPM}}
}
\examples{
### need to install package "limma"
### source("https://bioconductor.org/biocLite.R");biocLite("limma");
library(BPM);
BayPM(simUCEC,20,2);
}
\references{
Jianzhao Gao, Linghao Shen, and Xiaodan Fan,
Bayesian model for purity estimation using DNA methylation data.(submitted)
}
\author{
Jianzhao Gao(gaojz@nankai.edu.cn), Linghao Shen Xiaodan Fan (xfan@cuhk.edu.hk)
}
