% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUCEC.R
\docType{data}
\name{simUCEC}
\alias{simUCEC}
\title{Simulated data to illustrate datasets in packages}
\format{A matrix with 100 rows and 4 columns}
\description{
A dataset containing 100 gene and 4 smaples,
first two columns are tumor1 tumor2
last  two columns are normal1 normal2
\itemize{
\item x. the genes
\item y. two tumor samples; two normal samples;
}
}
\keyword{datasets}
